feat(csv): externalize the CSV crawler wizard to the plugin system
Migrates the CSV crawler's frontend wizard out of CrawlersPage.jsx into
tools/crawlers/csv/ (ConfigWizard.jsx + CrawlerMeta.js), matching the
self-contained pattern midPoint already uses. The backend half (crawler.json
+ Start-CSVCrawler.ps1) was already plugin-based; only the wizard needed to
move, and its existing onComplete()-based contract already matched the
plugin system's expectations with no logic changes needed.

Along the way:
- Unified CSV_SLOTS (frontend) and CSV_FILE_SLOTS (backend), which had
  already drifted — the backend's schema-download endpoint was silently
  missing the ContextMembers.csv slot. Both now read a single
  tools/crawlers/csv/csv-slots.json.
- Added a generic Summary.jsx plugin convention (discovered the same way
  CrawlerMeta.js is) so a crawler's configured-card panel doesn't need to be
  hardcoded per type in CrawlersPage.jsx. CSV's System/Type/Delimiter panel
  moved there with no visual change; any future externalized crawler can use
  the same mechanism.
- Widened vitest's include to also pick up co-located tests under
  tools/crawlers/**, and added the first tests for the CSV crawler's fuzzy
  filename matcher and its wizard.
- Removed csv from pr.yml's PENDING_MIGRATION list; fixed a stale
  onComplete(config) doc comment in tools/crawlers/CLAUDE.md that didn't
  match the actual bare onComplete() contract used by plugin wizards.

Verified with real Docker builds (frontend-build stage and full image) at
each step, a live container exercising the actual API (config create/list,
the now-fixed csv-schema endpoint), the full vitest + ESLint suite (0 lint
errors), and the crawler-manifest CI check's logic run locally.
