+++ tools/crawlers/csv/CrawlerMeta.js
@@ -1,0 +1,5 @@
+export default {
+  id: 'csv',
+  name: 'CSV Import',
+  description: 'Upload semicolon-delimited CSV files exported from Omada, SailPoint, or other IGA systems',
+};
