+++ tools/crawlers/csv/csv-slots.json
@@ -1,0 +1,12 @@
+[
+  { "key": "systems", "file": "Systems.csv", "label": "Systems", "required": false, "hint": "Optional. Columns: ExternalId, DisplayName, SystemType, Description" },
+  { "key": "contexts", "file": "Contexts.csv", "label": "Contexts", "required": false, "hint": "Optional. Columns: ExternalId, DisplayName, ContextType, TargetType, Description, ParentExternalId, SystemName, OwnerUserId" },
+  { "key": "context-members", "file": "ContextMembers.csv", "label": "Context Members", "required": false, "hint": "Optional. Columns: ContextExternalId, MemberExternalId, MemberType (Identity / Resource / Principal / System)." },
+  { "key": "resources", "file": "Resources.csv", "label": "Resources", "required": true, "hint": "Required. Columns: ExternalId, DisplayName, ResourceType, Description, SystemName, Enabled" },
+  { "key": "resourceRelationships", "file": "ResourceRelationships.csv", "label": "Resource Relationships", "required": false, "hint": "Optional. Columns: ParentExternalId, ChildExternalId, RelationshipType, SystemName" },
+  { "key": "users", "file": "Users.csv", "label": "Users", "required": true, "hint": "Required. Columns: ExternalId, DisplayName, Email, PrincipalType, JobTitle, Department, SystemName, Enabled" },
+  { "key": "assignments", "file": "Assignments.csv", "label": "Assignments", "required": true, "hint": "Required. Columns: ResourceExternalId, UserExternalId, AssignmentType, SystemName" },
+  { "key": "identities", "file": "Identities.csv", "label": "Identities", "required": false, "hint": "Optional. Columns: ExternalId, DisplayName, Email, EmployeeId, Department, JobTitle" },
+  { "key": "identityMembers", "file": "IdentityMembers.csv", "label": "Identity Members", "required": false, "hint": "Optional. Columns: IdentityExternalId, UserExternalId, AccountType" },
+  { "key": "certifications", "file": "Certifications.csv", "label": "Certifications", "required": false, "hint": "Optional. Columns: ExternalId, ResourceExternalId, UserDisplayName, Decision, ReviewerDisplayName, ReviewedDateTime" }
+]
